feat: changed dns footer link wording and functionality
