+++ src/components/Footer.jsx
@@ -16,7 +16,7 @@ import GooglePlayStoreButton from './GooglePlayStoreButton';
 import SocialIconLinks from './SocialIconLinks';
 import messages from './Footer.messages';
 import LanguageSelector from './LanguageSelector';
-import CCPADialog from './CCPADialog';
+import Cookies from 'js-cookie';
 
 ensureConfig([
   'LOGO_TRADEMARK_URL',
@@ -27,6 +27,8 @@ const EVENT_NAMES = {
 };
 
 const MARKETING_BASE_URL = 'https://edx.org';
+
+const CCPA_COOKIE_NAME = 'edx_do_not_sell';
 
 // Some MFEs (such as frontend-app-payment) don't want a footer to be shown.
 // This lets them hide it.
@@ -40,13 +42,10 @@ class Footer extends React.Component {
   constructor(props) {
     super(props);
     this.externalLinkClickHandler = this.externalLinkClickHandler.bind(this);
-    this.CCPADialogCloseHandler = this.CCPADialogCloseHandler.bind(this);
-    this.CCPADialogOpen = this.CCPADialogOpen.bind(this);
 
     this.state = {
       // Used for constructing the enterprise market link.
       utmSource: 'edx.org',
-      showCCPADialog: false,
     };
   }
 
@@ -79,18 +78,23 @@ class Footer extends React.Component {
     };
     sendTrackEvent(eventName, properties);
   }
+  
+  setDoNotSellCookie (value) {
+    const { host } = new URL(`${MARKETING_BASE_URL}`);
+    // Use global domain (`.edx.org`) without the subdomain
+    const hostParts = host.split('.');
+    const domain = hostParts.length > 2
+      ? `.${hostParts.slice(-2).join('.')}`
+      : host;
+    const cookieOptions = host.startsWith('localhost')
+      ? {}
+      : { domain, expires: 365 };
+    Cookies.set(CCPA_COOKIE_NAME, value, cookieOptions);
+  };
 
-  CCPADialogCloseHandler() {
-    this.setState({
-      showCCPADialog: false,
-    });
-  }
-
-  CCPADialogOpen() {
-    this.setState({
-      showCCPADialog: true,
-    });
-  }
+   toggleConsentModal () {
+    window.OneTrust?.ToggleInfoDisplay();
+  };
 
   render() {
     const {
@@ -198,7 +202,7 @@ class Footer extends React.Component {
               {
                 title: intl.formatMessage(messages['footer.legalLinks.doNotSellData']),
                 className: 'px-0 text-left text-decoration-none',
-                onClick: this.CCPADialogOpen,
+                onClick: () => {this.setDoNotSellCookie(true), this.toggleConsentModal()},
                 variant: 'link',
                 size: 'inline',
                 id: 'footer-dns-link',
@@ -264,11 +268,6 @@ class Footer extends React.Component {
             </p>
           </div>
         </div>
-        <CCPADialog
-          dialogIsOpen={this.state.showCCPADialog}
-          closeCallback={this.CCPADialogCloseHandler}
-          baseURL={`${MARKETING_BASE_URL}${localePrefix}`}
-        />
       </footer>
     );
   }
